short-circuit for-loop by using an explicit queue
Before this change, we were always looping through all the goals, even
if we already handled a goal and were just skipping every subsequent
goal.

Instead, this switches to an explicit mutable queue of remaining goals,
such that when a goal is handled, we can break out of the loop, append
all remaining goals at once, and go to the next iteration of the loop.

This seems to be improving some of the slower code, though I'm not fully
certain of it.  For instance, I seem to be getting 430s for
`./ZkLean/TwistShout/assert_lte.lean` where we used to get 510s.